requirements.txt

#mxnet
#pytorch
#tensorflow

numba
#pyculib

pynccl

zerorpc
msgpack==0.5.6
msgpack-numpy

psutil

six
